+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Possible log types:
 - `[fixed]` for any bug fixes.
 - `[security]` to invite users to upgrade in case of vulnerabilities.
 
-### Unreleased
+### v0.13.0 (2020-11-28)
 
 - [added] Added mandatory option `linker.platform` to `Drone.toml`
 - [added] Added option `probe.jlink.interface` to `Drone.toml`